+++ src/controllers/pgpool.js
@@ -1,0 +1,8 @@
+const { Pool } = require("pg");
+
+module.exports = new Pool({
+  connectionString: process.env.DATABASE_URL,
+  ssl: {
+    rejectUnauthorized: false,
+  },
+});
