+++ src/routes/index.js
@@ -1,22 +1,24 @@
 const { Router } = require("express");
 const router = Router();
 
-const controller = require("../controllers/index.controller");
+const userController = require("../controllers/user.controller");
+const imageController = require("../controllers/image.controller");
 
-router.get("/users", controller.getUsers);
-router.get("/users.compact", controller.getUsersCompact);
-router.get("/user/:username", controller.getUserByUsername);
-router.get("/user/id/:id", controller.getUserById);
-router.get("/user/email/:email", controller.getUserByEmail);
-router.post("/user.new", controller.createUser);
-router.delete("/user/id/:id", controller.deleteUser);
+router.get("/users", userController.getUsers);
+router.get("/users.compact", userController.getUsersCompact);
+router.get("/user/:username", userController.getUserByUsername);
+router.get("/user/id/:id", userController.getUserById);
+router.get("/user/email/:email", userController.getUserByEmail);
+router.patch("/user.new", userController.updateUserBio);
+router.post("/user.new", userController.createUser);
+router.delete("/user/id/:id", userController.deleteUser);
 
-router.get("/images", controller.getImages);
-router.get("/images.compact", controller.getImagesCompact);
-router.get("/image/:id", controller.getImageById);
-router.get("/image/:id/tags", controller.getImageTags);
-router.get("/user/:username/images", controller.getImagesByUsername);
-router.get("/user/id/:id/images", controller.getImagesByUserId);
-router.get("/user/email/:email/images", controller.getImagesByEmail);
+router.get("/images", imageController.getImages);
+router.get("/images.compact", imageController.getImagesCompact);
+router.get("/image/:id", imageController.getImageById);
+router.get("/image/:id/tags", imageController.getImageTags);
+router.get("/user/:username/images", imageController.getImagesByUsername);
+router.get("/user/id/:id/images", imageController.getImagesByUserId);
+router.get("/user/email/:email/images", imageController.getImagesByEmail);
 
 module.exports = router;
